hotfix: wp agent upload
